feature: preenchendo home com refeição consolidada

+++ app/src/main/java/com/example/controlealimentar/model/Refeicao.kt
@@ -4,8 +4,8 @@ data class Refeicao(
     val id: String,
     val nome: String,
     val horario: String = "00:00",
-    val caloria: Int = 0,
-    val proteina: Int = 0,
-    val carboidrato: Int = 0,
-    val gordura: Int = 0
+    val caloria: Double = 0.0,
+    val proteina: Double = 0.0,
+    val carboidrato: Double = 0.0,
+    val gordura: Double = 0.0
 )
